# Faster isolatedBufferGeometries()