--- a/src/analyzers/BufferGeometryMutator.js
+++ b/src/analyzers/BufferGeometryMutator.js
@@ -64,10 +64,8 @@ class BufferGeometryMutator {
         if (!this.findNeighbors()) {
             return null;
         }
-
         this.removeDegenerates(this.range(this.positions.length/9));
         this.deleteDegenerates();
-
         return this;
     }