… test/BufferGeometryMutatorTest.js
… test/isolateGeometries.js