This improves the speed of isolatedBufferGeometries() to even better than it was before the previous merge.  Also, support for colors is added.  To see it work, use `npm run dev`.

To measure speed, I used Chrome and opened the Developer Tools.  I use the Javascript Profiler, started it, then split the example, then stopped it, and looked at the "Total Time" of isolatedGeometries.  I did it 3 times.

First, the original code at commit `535e9f5` was 1002.8ms, 992.1ms, 991.5ms.

The code before this PR is 2012.5ms, 1997.1ms, and 2021.8ms.

The problem was that I was splitting the geometries by first converted them into a bunch of `BufferGeometryMutators` and then converting each into a `BufferGeometry`.  That extra conversion in the middle slowed it down.  So I now convert directly to BufferGeometries, which gave a good improvement.  I also added in code to support colors so that the example will work better.  Also some other speed improvements.

With this PR, it's now 851.5ms, 865.0ms, 920.5ms.