@eyal0 2017-11-18
Thanks for checking the speed, by the way.  That test is a good one because there are many shapes in there and they are each complicated.